+++ frontend/openchat-agent/src/services/userIndex/userIndex.client.ts
@@ -689,8 +689,8 @@ export class UserIndexClient extends CandidService {
             "update_bot",
             {
                 bot_id: principalStringToBytes(id),
-                principal: mapOptional(principal, principalStringToBytes),
                 owner: mapOptional(ownerId, principalStringToBytes),
+                principal: mapOptional(principal, principalStringToBytes),
                 avatar:
                     mapOptional(avatarUrl, (url) => ({
                         SetToSome: url,